Add comment

+++ line_plot.py
@@ -1,7 +1,8 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-train = np.loadtxt('click.csv', delimiter=',', skiprows=1)
+# 学習データの読み込み
+train = np.loadtxt('click.csv', delimiter=',', dtype='int', skiprows=1)
 train_x = train[:,0]
 train_y = train[:,1]
 
@@ -13,6 +14,7 @@ def standardize(x):
 
 train_z = standardize(train_x)
 
+# パラメータを初期化
 theta0 = np.random.rand()
 theta1 = np.random.rand()
 
@@ -24,9 +26,6 @@ def f(x):
 def E(x, y):
   return 0.5 * np.sum((y - f(x)) ** 2)
 
-
-# 最急降下法
-
 # 学習率
 ETA = 1e-3
 
@@ -36,24 +35,28 @@ diff = 1
 # 更新回数
 count = 0
 
-# 学習を繰り返す
+# 誤差の差分んが0.01以下になるまでパラメータ更新を繰り返す
 error = E(train_z, train_y)
 while diff > 1e-2:
   # 更新結果を一時変数に保存
   tmp0 = theta0 - ETA * np.sum((f(train_z) - train_y))
   tmp1 = theta1 - ETA * np.sum((f(train_z) - train_y) * train_z)
+
   # パラメータ更新
   theta0 = tmp0
   theta1 = tmp1
-  # 前回の誤差
+
+  # 前回の誤差との差分計算
   current_error = E(train_z, train_y)
   diff = error - current_error
   error = current_error
+
   # ログの出力
   count += 1
   log = '{}回目 : theta0 = {:.3f}, theta1 = {:.3f}, 差分 = {:.4f}'
   print(log.format(count, theta0, theta1, diff))
 
+# プロットして確認
 x = np.linspace(-3, 3, 100)
 plt.plot(train_z, train_y, 'o')
 plt.plot(x, f(x))
